move out CommonAncestor from RdObject
